Rename ASTBaseNode to AbstractStatementNode
