+++ src/main/java/com/merkrafter/representation/ast/AbstractStatementNode.java
@@ -4,16 +4,15 @@ import java.util.LinkedList;
 import java.util.List;
 
 /****
- * This is the base node of all node types that an Abstract Syntax Tree
- * may be made of. The AST represents the program's structure.
+ * This is the base node of all statement node types that an Abstract Syntax Tree
+ * may be made of.
+ * <p>
+ * A Statement actually is a linked list of multiple statements that are executed in sequence.
  *
  * @since v0.3.0
  * @author merkrafter
  ***************************************************************/
-public abstract class ASTBaseNode implements Statement {
-    // CONSTANTS
-    //==============================================================
-
+public abstract class AbstractStatementNode implements Statement {
     // ATTRIBUTES
     //==============================================================
     /**
@@ -27,7 +26,7 @@ public abstract class ASTBaseNode implements Statement {
     /****
      * Creates a new node without a next node assigned.
      ***************************************************************/
-    public ASTBaseNode() {
+    public AbstractStatementNode() {
     }
 
     // GETTER
@@ -52,6 +51,12 @@ public abstract class ASTBaseNode implements Statement {
         this.next = next;
     }
 
+    /**
+     * This is a utility method that can be used by nodes to check their children for errors.
+     *
+     * @param nodes child node(s) that should be tested
+     * @return a list of all errors that can be found in this subtree
+     */
     static List<String> collectErrorsFrom(final AbstractSyntaxTree... nodes) {
         final List<String> errors = new LinkedList<>();
         for (final AbstractSyntaxTree node : nodes) {
