+++ src/main/java/com/merkrafter/representation/ast/ProcedureCallNode.java
@@ -11,7 +11,7 @@ import java.util.List;
  * @since v0.3.0
  * @author merkrafter
  ***************************************************************/
-public class ProcedureCallNode extends ASTBaseNode implements Expression {
+public class ProcedureCallNode extends AbstractStatementNode implements Expression {
     // ATTRIBUTES
     //==============================================================
     private final ProcedureDescription procedure;
